+++ .pmci/netbsd.sh
@@ -4,6 +4,6 @@ set -ex
 export GOPATH=/tmp/go
 mkdir -p /tmp/go/src/github.com/winfsp
 cp -R /tmp/repo/cgofuse /tmp/go/src/github.com/winfsp
-cd /tmp/go/src/github.com/winfsp/cgofuse
+cd /tmp/go/src/github.com/AmericanBinary/cgofuse
 go build -v ./...
 # go test -v ./fuse -run 'TestUnmount|TestSignal'